feat : 격자자료를 파싱하는 parseGridData 메소드가 List<List<Float>>이 아닌 List<List<Integer>>를 반환하도록 수정.

+++ src/main/java/com/github/yun531/climate/weatherApi/WeatherApiClient.java
@@ -25,7 +25,7 @@ public class WeatherApiClient {
         this.restClient = RestClient.create();
     }
 
-    public List<List<Float>> RequestShortTermGridForecast(int hoursToTargetTime, ForecastCategory forecastVar) throws URISyntaxException {
+    public List<List<Integer>> RequestShortTermGridForecast(int hoursToTargetTime, ForecastCategory forecastVar) throws URISyntaxException {
         LocalDateTime nowDateTime = LocalDateTime.now();
         LocalDateTime targetTime = nowDateTime.plusHours(hoursToTargetTime);
